Move the additional edge link accessors and manipulators to a new package private header file
