+++ Makefile.am
@@ -50,6 +50,7 @@ nobase_pkginclude_HEADERS = \
 	c/graphLib/graph.h \
 	c/graphLib/graphStructures.h \
 	c/graphLib/graphDFSUtils.h \
+	c/graphLib/graphUtils.private.h \
 	c/graphLib/extensionSystem/graphExtensions.h \
 	c/graphLib/extensionSystem/graphExtensions.private.h \
 	c/graphLib/extensionSystem/graphFunctionTable.h \
